Broke the Implicits out of the raster package object, and put them into Implicits.scala
